+++ template/src/main.tsx
@@ -1,5 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import 'matcha-oat-design-system/tokens.css'
+import 'matcha-oat-design-system/fonts.css'
 import './index.css'
 import App from './App.tsx'